requirements.txt
appnope==0.1.4
asttokens==2.4.1
blinker==1.8.2
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
comm==0.2.2
DateTime==5.5
debugpy==1.8.2
decorator==5.1.1
dnspython==2.6.1
exceptiongroup==1.2.1
executing==2.0.1
Flask==3.0.3
greenlet==3.0.3
idna==3.7
ipykernel==6.29.5
ipython==8.26.0
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
jupyter_client==8.6.2
jupyter_core==5.7.2
MarkupSafe==2.1.5
matplotlib-inline==0.1.7
nest-asyncio==1.6.0
numpy==2.0.0
packaging==24.1
pandas==2.2.2
parso==0.8.4
pexpect==4.9.0
platformdirs==4.2.2
prompt_toolkit==3.0.47
psutil==6.0.0
psycopg2==2.9.9
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.18.0
pymongo==4.8.0
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
pyzmq==26.0.3
requests==2.32.3
six==1.16.0
SQLAlchemy==2.0.31
SQLAlchemy-Utils==0.41.2
stack-data==0.6.3
tornado==6.4.1
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
wcwidth==0.2.13
Werkzeug==3.0.3
xmltodict==0.13.0
zope.interface==6.4.post2
python-dotenv==1.0.1